feat: seperate login and eventEdit to different component
